Add cmd line parameters for early logging
Added the following command-line parameters to `firecracker`:
`--log-path`, `--level`, `--show-level` and `--show-log-origin`
that can be used for configuring the Logger when starting the
process and without needing to send an API request. When using
this method for configuration, only `--log-path` is mandatory.
The value for this parameter should be the destination for the
fifo/file. This resource should already exist at the specified
path.

Signed-off-by: Laura Loghin <lauralg@amazon.com>