Decouple logger and metrics functionality
Added a new API call, `PUT /metrics`, for configuring the metrics
system. Now, the logger and metrics system are decoupled at both API
level and logger crate level in order to obtain more flexibility.
This way the users are not forced anymore to set also a metrics pipe
even if they are interested only in the logging functionality.

Signed-off-by: Laura Loghin <lauralg@amazon.com>